+++ src/account/BiconomySmartAccountV2.ts
@@ -137,6 +137,7 @@ export class BiconomySmartAccountV2 extends BaseSmartContractAccount {
       ...biconomySmartAccountConfig,
       chain:
         biconomySmartAccountConfig.viemChain ??
+        biconomySmartAccountConfig.customChain ??
         getChain(biconomySmartAccountConfig.chainId),
       rpcClient:
         biconomySmartAccountConfig.rpcUrl ||
@@ -199,6 +200,7 @@ export class BiconomySmartAccountV2 extends BaseSmartContractAccount {
     this.provider = createPublicClient({
       chain:
         biconomySmartAccountConfig.viemChain ??
+        biconomySmartAccountConfig.customChain ??
         getChain(biconomySmartAccountConfig.chainId),
       transport: http(
         biconomySmartAccountConfig.rpcUrl ||
@@ -283,12 +285,17 @@ export class BiconomySmartAccountV2 extends BaseSmartContractAccount {
     if (!chainId) {
       throw new Error("chainId required")
     }
+
     const bundler: IBundler =
       biconomySmartAccountConfig.bundler ??
       new Bundler({
-        // biome-ignore lint/style/noNonNullAssertion: <explanation>
+        // biome-ignore lint/style/noNonNullAssertion: always required
         bundlerUrl: biconomySmartAccountConfig.bundlerUrl!,
-        chainId
+        chainId,
+        customChain:
+          biconomySmartAccountConfig.viemChain ??
+          biconomySmartAccountConfig.customChain ??
+          getChain(chainId)
       })
     let defaultValidationModule =
       biconomySmartAccountConfig.defaultValidationModule
@@ -882,7 +889,7 @@ export class BiconomySmartAccountV2 extends BaseSmartContractAccount {
 
   // dummy signature depends on the validation module supplied.
   async getDummySignatures(_params?: ModuleInfo): Promise<Hex> {
-    const params = { ..._params, ...(this.sessionData ? this.sessionData : {}) }
+    const params = { ...(this.sessionData ? this.sessionData : {}), ..._params }
     this.isActiveValidationModuleDefined()
     return (await this.activeValidationModule.getDummySignature(params)) as Hex
   }
@@ -913,7 +920,7 @@ export class BiconomySmartAccountV2 extends BaseSmartContractAccount {
     userOp: Partial<UserOperationStruct>,
     _params?: SendUserOpParams
   ): Promise<UserOperationStruct> {
-    const params = { ..._params, ...(this.sessionData ? this.sessionData : {}) }
+    const params = { ...(this.sessionData ? this.sessionData : {}), ..._params }
 
     this.isActiveValidationModuleDefined()
     const requiredFields: UserOperationKey[] = [
@@ -1526,6 +1533,8 @@ export class BiconomySmartAccountV2 extends BaseSmartContractAccount {
 
   /**
    * Builds a user operation
+   *
+   * This method will also simulate the validation and execution of the user operation, telling the user if the user operation will be successful or not.
    *
    * - Docs: https://docs.biconomy.io/Account/transactions/userpaid#build-useroperation
    *